refactor: rename creds templates to credentials
